% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Marathon}
\alias{Marathon}
\title{Marathon Runners}
\format{A data frame with 200 observations on the following 9 variables.
\describe{
\item{id}{ID number of the runner}
\item{first_name}{Runner's first name}
\item{last_name}{Runner's last name}
\item{sex}{Sex of the runner:  "F" or "M"}
\item{dob}{Date of birth of the runner}
\item{team}{ID of the team (if any) that the runner belongs to}
\item{time}{Runner's finishing time, as hours, minutes and seconds}
\item{seconds}{Runner's finishing time in seconds}
\item{team_name}{full name of the team (if any) the runner belongs to}
}}
\source{
Hypothetical data.
}
\description{
A table produced by joining some other tables
in this package.
}
\keyword{datasets}
