% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Racers}
\alias{Racers}
\title{Marathon Runners}
\format{A data frame with 200 observations on the following 6 variables.
\describe{
\item{id}{ID number of the runner}
\item{first_name}{Runner's first name}
\item{last_name}{Runner's last name}
\item{sex}{Sex of the runner:  "F" or "M"}
\item{dob}{Date of birth of the runner}
\item{team}{ID of the team (if any) that the runner belongs to}
}}
\source{
Hypothetical data.
}
\description{
Runners in a hypothetical marathon.
}
\keyword{datasets}
