% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Results}
\alias{Results}
\title{Marathon Results}
\format{A data frame with 200 observations on the following 3 variables.
\describe{
\item{id}{ID number of the runner}
\item{time}{Runner's finishing time, as hours, minutes and seconds}
\item{seconds}{Runner's finishing time in seconds}
}}
\source{
Hypothetical data.
}
\description{
Results of a hypothetical marathon.
}
\keyword{datasets}
