% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teams}
\alias{teams}
\title{Marathon Teams}
\format{A data frame with 15 observations on the following 2 variables.
\describe{
\item{teamID}{ID of the team}
\item{team_name}{full name of the team}
}}
\source{
Hypothetical data.
}
\description{
Teams in a hypothetical marathon.
}
\keyword{datasets}
